packages/qa/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "setuptools_scm>=8"]

[dependency-groups]
dev = ["ether0.qa[dev]"]

[project]
authors = [
    {email = "hello@futurehouse.org", name = "FutureHouse technical staff"},
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Chemistry",
]
dependencies = [
    "datasets",
    "fhaviary[litqa]>=0.19",  # Pin for Python 3.13 compatibility
    "fhlmi>=0.26",  # Pin for Python 3.13 compatibility
    "futurehouse-client>=0.3.20.dev63",  # Pin for WorldModel primitive and FUTUREHOUSE_API_KEY
    "ldp>=0.26",  # Pin for Python 3.13 compatibility
    "litellm>=1.60.0",  # Pin for LoggingCallbackManager refactor
    "paper-qa @ git+https://github.com/Future-House/paper-qa.git@working-with-py-typed",  # For py.typed support
    "paperqa-server[litqa_v2] @ git+ssh://git@github.com/Future-House/paperqa-server.git@world-models",  # For Andrew's WorldModel env
    "pydantic>=2",  # Pin to keep recent
    "tqdm",
]
description = "Question-answer code based on the ether0 source code"
dynamic = ["version"]
name = "ether0.qa"
readme = "README.md"
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "ether0.qa[typing]",
    "pytest",
    "pytest-asyncio",
    "pytest-sugar",
    "pytest-timer[colorama]",
    "pytest-xdist",
]
typing = [
    "anyio",
    "types-tqdm",
]

[tool.ruff]
extend = "../../pyproject.toml"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
root = "../.."
